[Decomp][Player] Cohesion split: extract regen methods to PlayerRegen.cpp
